Issue #4220: Split and Organize Checkstyle inputs by Test in the blocks package: RightCurly
