+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/blocks/rightcurly/InputRightCurly.java
@@ -2,7 +2,7 @@
 // Test case file for checkstyle.
 // Created: 2001
 ////////////////////////////////////////////////////////////////////////////////
-package com.puppycrawl.tools.checkstyle.checks.blocks;
+package com.puppycrawl.tools.checkstyle.checks.blocks.rightcurly;
 
 class InputRightCurly
 {
