+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/blocks/rightcurly/InputRightCurlyNameForOneLiners.java
@@ -2,7 +2,7 @@
 // Test case file for checkstyle.
 // Created: 2015
 ////////////////////////////////////////////////////////////////////////////////
-package com.puppycrawl.tools.checkstyle.checks.blocks;
+package com.puppycrawl.tools.checkstyle.checks.blocks.rightcurly;
 
 /**
  * Test case for RightCurly with option SAME to omit oneliners
